I-14: Changed int32_t system event values to int64

+++ include/wui/window/window.hpp
@@ -283,7 +283,7 @@ private:
     void draw_border(graphic &gr);
 
     void send_internal(internal_event_type type, int32_t x, int32_t y);
-    void send_system(system_event_type type, int32_t x, int32_t y);
+    void send_system(system_event_type type, uint64_t x, uint64_t y);
 };
 
 }
